+++ CHANGELOG.md
@@ -17,6 +17,10 @@ Here is a template for new release sections
 -
 ```
 
+## [Unreleased]
+### Fixed
+- SPARQL queries
+
 ## [0.20.3] - 2024-03-13
 ### Fixed
 - multiple versions error when merging FKs into process